update: public modal
